chore: adjust layout
