+++ src/pages/_app.tsx
@@ -4,6 +4,8 @@ import '@/styles/globals.css';
 // !STARTERCONF This is for demo purposes, remove @/styles/colors.css import immediately
 import '@/styles/colors.css';
 
+import Layout from '@/components/layout/Layout';
+
 /**
  * !STARTERCONF info
  * ? `Layout` component is called in every page using `np` snippets. If you have consistent layout across all page, you can add it here too
@@ -12,7 +14,9 @@ import '@/styles/colors.css';
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <div id='root' className='min-h-screen bg-gray-100'>
-      <Component {...pageProps} />
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
     </div>
   );
 }
